chore: move cell editor widgets to its own folder
