+++ frontend/appflowy_flutter/lib/plugins/database/widgets/cell_editor/select_option_editor.dart
@@ -12,12 +12,12 @@ import 'package:flowy_infra_ui/style_widget/hover.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../grid/presentation/layout/sizes.dart';
-import '../../../../grid/presentation/widgets/common/type_option_separator.dart';
-import '../../../../grid/presentation/widgets/header/type_option/select/select_option_editor.dart';
+import '../../grid/presentation/layout/sizes.dart';
+import '../../grid/presentation/widgets/common/type_option_separator.dart';
+import '../../grid/presentation/widgets/header/type_option/select/select_option_editor.dart';
 import 'extension.dart';
-import '../../../../application/cell/bloc/select_option_editor_bloc.dart';
-import 'text_field.dart';
+import '../../application/cell/bloc/select_option_editor_bloc.dart';
+import 'select_option_text_field.dart';
 
 const double _editorPanelWidth = 300;
