--- frontend/appflowy_flutter/lib/plugins/database/widgets/row/cells/url_cell/url_cell_editor.dart
@@ -1,106 +1,0 @@
-import 'package:appflowy/plugins/database/application/cell/cell_controller_builder.dart';
-import 'package:flutter/material.dart';
-import 'dart:async';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../../../../application/cell/bloc/url_cell_editor_bloc.dart';
-
-class URLCellEditor extends StatefulWidget {
-  const URLCellEditor({
-    super.key,
-    required this.cellController,
-    required this.onExit,
-  });
-
-  final URLCellController cellController;
-  final VoidCallback onExit;
-
-  @override
-  State<URLCellEditor> createState() => _URLCellEditorState();
-}
-
-class _URLCellEditorState extends State<URLCellEditor> {
-  late URLCellEditorBloc _cellBloc;
-  late TextEditingController _controller;
-
-  @override
-  void initState() {
-    _cellBloc = URLCellEditorBloc(cellController: widget.cellController);
-    _cellBloc.add(const URLCellEditorEvent.initial());
-    _controller = TextEditingController(text: _cellBloc.state.content);
-
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider.value(
-      value: _cellBloc,
-      child: BlocListener<URLCellEditorBloc, URLCellEditorState>(
-        listener: (context, state) {
-          if (_controller.text != state.content) {
-            _controller.text = state.content;
-          }
-
-          if (state.isFinishEditing) {
-            widget.onExit();
-          }
-        },
-        child: TextField(
-          autofocus: true,
-          controller: _controller,
-          onSubmitted: (value) => focusChanged(),
-          onEditingComplete: () => focusChanged(),
-          style: Theme.of(context).textTheme.bodyMedium,
-          decoration: const InputDecoration(
-            contentPadding: EdgeInsets.zero,
-            border: InputBorder.none,
-            hintText: "",
-            isDense: true,
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  Future<void> dispose() async {
-    _cellBloc.close();
-    super.dispose();
-  }
-
-  void focusChanged() {
-    if (mounted) {
-      if (_cellBloc.isClosed == false &&
-          _controller.text != _cellBloc.state.content) {
-        _cellBloc.add(URLCellEditorEvent.updateText(_controller.text));
-      }
-    }
-  }
-}
-
-class URLEditorPopover extends StatelessWidget {
-  const URLEditorPopover({
-    super.key,
-    required this.cellController,
-    required this.onExit,
-  });
-
-  final URLCellController cellController;
-  final VoidCallback onExit;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(4),
-      ),
-      padding: const EdgeInsets.all(12),
-      child: URLCellEditor(
-        cellController: cellController,
-        onExit: onExit,
-      ),
-    );
-  }
-}
